+++ src/utils/dex/functions.js
@@ -1,18 +1,23 @@
 import { PersonalTable } from '../../../__gamedata';
-import { getPokemonName, getAbilityString, getTechMachineLearnset, getWeight, getHeight, getTypeName } from './';
+import { getPokemonName } from './name';
+import { getAbilityString } from './ability';
+import { getWeight, getHeight } from './details';
+import { getTypeName } from './types';
 
 //BDSP does not stick to the same structure when working with forms, thus this map is necessary.
-const FORM_MAP = PersonalTable.Personal.reduce((formMap, currentPokemon) => {
+const FORM_MAP = PersonalTable.Personal.reduce(createFormMap, {});
+
+function createFormMap(formMap, currentPokemon) {
   if (!Array.isArray(formMap[currentPokemon.monsno])) {
     formMap[currentPokemon.monsno] = [];
   }
 
   formMap[currentPokemon.monsno].push(currentPokemon.id);
   return formMap;
-}, {});
+}
 
 function getPokemonIdFromFormMap(monsNo = 0, formNo = 0) {
-  return FORM_MAP[monsNo][formNo];
+  return FORM_MAP[monsNo]?.[formNo] ?? undefined;
 }
 
 function getGender(sex) {
@@ -22,12 +27,10 @@ function getGender(sex) {
   return null;
 }
 
-function parseTmLearnsetSection(decimal) {
-  return (decimal >>> 0).toString(2).split('').reverse().join('').padStart(32, 0);
-}
-
-function getImage(monsno = 0) {
-  return `/img/pm${monsno.toString().padStart(4, 0) ?? '0000'}_00_00_00_L.webp`;
+function getImage(monsno = 0, formIndex = 0) {
+  const paddedMonsno = monsno.toString().padStart(4, 0);
+  const paddedFormIndex = formIndex.toString().padStart(2, 0);
+  return `/img/pm${paddedMonsno}_${paddedFormIndex}_00_00_L.webp`;
 }
 
 function formatBaseStats(p) {
@@ -47,38 +50,6 @@ function getPokemonIdFromMonsNoAndForm(monsno, formno) {
   return PersonalTable.Personal.find((e) => e.monsno === monsno && FORM_MAP[e.monsno][formno] === e.id)?.id;
 }
 
-function getPokemonInfo(monsno = 0) {
-  const p = PersonalTable.Personal[monsno];
-  return {
-    monsno: monsno,
-    name: getPokemonName(monsno),
-    ability1: getAbilityString(p.tokusei1),
-    ability2: getAbilityString(p.tokusei2),
-    abilityH: getAbilityString(p.tokusei3),
-    tmLearnset: getTechMachineLearnset(p.machine1, p.machine2, p.machine3, p.machine4),
-    prettyBaseStats: formatBaseStats(p),
-    baseStats: {
-      hp: p.basic_hp,
-      atk: p.basic_atk,
-      def: p.basic_def,
-      spa: p.basic_spatk,
-      spd: p.basic_spdef,
-      spe: p.basic_agi,
-    },
-    baseStatsTotal: p.basic_hp + p.basic_atk + p.basic_def + p.basic_spatk + p.basic_spdef + p.basic_agi,
-    weight: getWeight(monsno),
-    height: getHeight(monsno),
-    grassKnotPower: getGrassKnotPower(getWeight(monsno)),
-    type1: getTypeName(p.type1),
-    type2: getTypeName(p.type2),
-    imageSrc: getImage(monsno),
-  };
-}
-
-function getPokemonFormId(monsno = 0, id) {
-  return FORM_MAP[monsno].findIndex((e) => e === id);
-}
-
 export {
   FORM_MAP,
   getPokemonIdFromFormMap,
@@ -86,8 +57,6 @@ export {
   getGrassKnotPower,
   getImage,
   formatBaseStats,
-  parseTmLearnsetSection,
   getPokemonIdFromMonsNoAndForm,
-  getPokemonInfo,
-  getPokemonFormId,
+  createFormMap,
 };
