+++ src/utils/dex/index.js
@@ -9,6 +9,7 @@ import {
   getTechMachineLearnset,
   getMoveProperties,
   getPokemonLearnset,
+  parseTmLearnsetSection,
 } from './moves';
 import {
   getPokemonMonsNoFromName,
@@ -17,6 +18,7 @@ import {
   getPokemonName,
   getPokemonIdFromName,
   getPokemonNames,
+  getPokemonFormId,
 } from './name';
 import { getNatureId, getNatureName } from './nature';
 import { getTypeName, getTypes } from './types';
@@ -28,16 +30,43 @@ import {
   getGrassKnotPower,
   getImage,
   formatBaseStats,
-  parseTmLearnsetSection,
   getPokemonIdFromMonsNoAndForm,
-  getPokemonInfo,
-  getPokemonFormId,
 } from './functions';
+
+import { PersonalTable } from '../../../__gamedata';
 
 const POKEMON_MOVE_LEVEL_TYPE = {
   EGG: 'egg',
   TM: 'tm',
 };
+
+function getPokemonInfo(monsno = 0) {
+  const p = PersonalTable.Personal[monsno];
+  return {
+    monsno: monsno,
+    name: getPokemonName(monsno),
+    ability1: getAbilityString(p.tokusei1),
+    ability2: getAbilityString(p.tokusei2),
+    abilityH: getAbilityString(p.tokusei3),
+    tmLearnset: getTechMachineLearnset(p.machine1, p.machine2, p.machine3, p.machine4),
+    prettyBaseStats: formatBaseStats(p),
+    baseStats: {
+      hp: p.basic_hp,
+      atk: p.basic_atk,
+      def: p.basic_def,
+      spa: p.basic_spatk,
+      spd: p.basic_spdef,
+      spe: p.basic_agi,
+    },
+    baseStatsTotal: p.basic_hp + p.basic_atk + p.basic_def + p.basic_spatk + p.basic_spdef + p.basic_agi,
+    weight: getWeight(monsno),
+    height: getHeight(monsno),
+    grassKnotPower: getGrassKnotPower(getWeight(monsno)),
+    type1: getTypeName(p.type1),
+    type2: getTypeName(p.type2),
+    imageSrc: getImage(monsno),
+  };
+}
 
 export {
   FORM_MAP,
